add/delete friend/group button changed correctly

+++ src/Group_Timeline.java
@@ -6,13 +6,14 @@ import java.awt.event.ActionListener;
 import javax.swing.ImageIcon;
 import javax.swing.JButton;
 import javax.swing.JFrame;
+import javax.swing.JPanel;
 import javax.swing.JScrollPane;
 import javax.swing.ScrollPaneConstants;
-import javax.swing.JPanel;
 
 public class Group_Timeline {
 
 	private JFrame frame;
+	private JButton isMemberButton;
 	private JPanel panel;
 	private JButton addgroup;
 	private JButton deletegroup;
@@ -38,24 +39,11 @@ public class Group_Timeline {
 		frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
 		frame.getContentPane().setLayout(null);
 		
-		addgroup = new JButton("Add group");
-		addgroup.setFont(new Font("Arial", Font.PLAIN, 16));
-		addgroup.setBounds(57, 13, 119, 31);
-		frame.getContentPane().add(addgroup);
-		
-		deletegroup = new JButton("Delete group");
-		deletegroup.setFont(new Font("Arial", Font.PLAIN, 16));
-		deletegroup.setBounds(237, 13, 137, 31);
-		frame.getContentPane().add(deletegroup);
-		
-        addgroup.addActionListener(new addgroupListener());
-        deletegroup.addActionListener(new deletegroupListener());
-
-//		btnMorePosts = new JButton("More Posts");
-//		btnMorePosts.setBounds(204, 333, 143, 25);
-//		btnMorePosts.setFont(new Font("Arial", Font.PLAIN, 16));
-//		btnMorePosts.addActionListener(new NextPostsListener());
-//		frame.getContentPane().add(btnMorePosts);
+		isMemberButton = new JButton();
+		isMemberButton.setFont(new Font("Arial", Font.PLAIN, 16));
+		isMemberButton.setBounds(103, 13, 149, 31);
+		isMemberButton.addActionListener(new isMemberButtonActionListener());
+		frame.getContentPane().add(isMemberButton);
 		
 		JButton btnBack = new JButton("Back");
 		btnBack.setBounds(535, 333, 97, 25);
@@ -66,76 +54,42 @@ public class Group_Timeline {
 		JButton Members_List = new JButton("Members");
 		Members_List.addActionListener(new MembersListener());
 		Members_List.setFont(new Font("Arial", Font.PLAIN, 16));
-		Members_List.setBounds(437, 13, 137, 31);
+		Members_List.setBounds(356, 13, 137, 31);
 		frame.getContentPane().add(Members_List);
 		
 		panel = new Post_View(u,g);
-//		panel.setBounds(12, 66, 620, 262);
 		scrollpane = new JScrollPane(panel,ScrollPaneConstants.VERTICAL_SCROLLBAR_AS_NEEDED, ScrollPaneConstants.HORIZONTAL_SCROLLBAR_AS_NEEDED);
 		scrollpane.setBounds(12, 66, 620, 262);
 		scrollpane.setBorder(null);
-		
-		  if(g.isMember(u)){
-	        	addgroup.setEnabled(false);
-	        	frame.getContentPane().add(scrollpane);
-//	        	frame.getContentPane().add(panel);
-	        }
-	        else{
-	        	deletegroup.setEnabled(false);
-//	        	btnMorePosts.setEnabled(false);
-	        }
-		
-	}
-	
-	class addgroupListener implements ActionListener{
-		public void actionPerformed(ActionEvent e){
-			g.addMember(u);			
-			addgroup.setEnabled(false);
-			deletegroup.setEnabled(true);
-//			btnMorePosts.setEnabled(true);
-	        DataBase.save();
-	        panel = new Post_View(u,g);
-//			panel.setBounds(12, 59, 620, 261);
+		if(g.isMember(u)){
+			isMemberButton.setText("Delete Group");
 			frame.getContentPane().add(scrollpane);
-			frame.repaint();
-			frame.revalidate();
-	        
 		}
-			
+		else 
+			isMemberButton.setText("Add Group");
 	}
 	
-	class deletegroupListener implements ActionListener{
+	class isMemberButtonActionListener implements ActionListener{
 		public void actionPerformed(ActionEvent e){
+			if(isMemberButton.getText().equals("Add Group")){
+				g.addMember(u);
+				DataBase.save();
+				isMemberButton.setText("Delete Group");
+		        panel = new Post_View(u,g);
+				frame.getContentPane().add(scrollpane);
+				frame.repaint();
+				frame.revalidate();
+			}
+			else if(isMemberButton.getText().equals("Delete Group")){
 				u.deleteFromGroup(g);
-				addgroup.setEnabled(true);
-				deletegroup.setEnabled(false);
-	        	DataBase.save();
-	        	
-//	        	if(btnMorePosts.isEnabled())
-//	        		frame.getContentPane().remove(panel);
-//	        	else{
-	        		frame.getContentPane().remove(scrollpane);
-//	        	}
-	        		
-//	        	btnMorePosts.setEnabled(false);
+				DataBase.save();
+				isMemberButton.setText("Add Group");
+				frame.getContentPane().remove(scrollpane);
 	        	frame.getContentPane().revalidate();
 	        	frame.getContentPane().repaint();
+			}
 		}
-			
 	}
-	
-//	class NextPostsListener implements ActionListener{
-//		public void actionPerformed(ActionEvent e){
-//			frame.getContentPane().remove(panel);
-//			scrollpane = new JScrollPane(panel);
-//			scrollpane.setBounds(12, 66, 601, 262);
-//			scrollpane.setBorder(null);
-//			frame.getContentPane().add(scrollpane);
-//			btnMorePosts.setEnabled(false);
-//        	frame.getContentPane().repaint();
-//        	frame.getContentPane().revalidate();
-//		}
-//	}
 	
 	class BackListener implements ActionListener{
 		public void actionPerformed(ActionEvent e){
@@ -147,7 +101,7 @@ public class Group_Timeline {
 	
 	class MembersListener implements ActionListener{
 		public void actionPerformed(ActionEvent e){
-			DisplayLists.createAndShowGUI(u,g.members ,null);
+			DisplayLists.createAndShowGUI(u,g.members ,null,frame);
 			
 		}
 	}
